backend/requirements.txt
django-extensions
Django==2.0.2
psycopg2
djangorestframework
requests
django-cors-headers==2.4.0
